[refactor] entity 기반 구조로 변경
- goal, template로 나눠 각각의 패키지로 변경
