+++ backend/src/main/java/com/hotpack/krocs/domain/goals/controller/GoalController.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.goals.controller;
+
+public class GoalController {
+}
