+++ backend/src/main/java/com/hotpack/krocs/domain/goals/dto/GoalDTO.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.goals.dto;
+
+public class GoalDTO {
+}
