+++ backend/src/main/java/com/hotpack/krocs/domain/goals/exception/GoalExceptionType.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.goals.exception;
+
+public class GoalExceptionType {
+}
