+++ backend/src/main/java/com/hotpack/krocs/domain/goals/repository/SubGoalRepository.java
@@ -1,0 +1,10 @@
+package com.hotpack.krocs.domain.goals.repository;
+
+import com.hotpack.krocs.domain.goals.domain.SubGoal;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SubGoalRepository extends JpaRepository<SubGoal, Long> {
+
+}
