+++ backend/src/main/java/com/hotpack/krocs/domain/goals/service/GoalService.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.goals.service;
+
+public interface GoalService {
+}
