+++ backend/src/main/java/com/hotpack/krocs/domain/goals/service/GoalServiceImpl.java
@@ -1,0 +1,7 @@
+package com.hotpack.krocs.domain.goals.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class GoalServiceImpl {
+}
