+++ backend/src/main/java/com/hotpack/krocs/domain/templates/controller/TemplateController.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.templates.controller;
+
+public class TemplateController {
+}
