+++ backend/src/main/java/com/hotpack/krocs/domain/templates/domain/SubTemplate.java
@@ -1,4 +1,4 @@
-package com.hotpack.krocs.common.entity.goals;
+package com.hotpack.krocs.domain.templates.domain;
 
 import com.hotpack.krocs.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
@@ -8,21 +8,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "sub_goals")
+@Table(name = "sub_templates")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class SubGoal extends BaseTimeEntity {
+public class SubTemplate extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "sub_goal_id")
-    private Long subGoalId;
+    @Column(name = "sub_template_id")
+    private Long subTemplateId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "goal_id", nullable = false)
-    private Goal goal;
+    @JoinColumn(name = "template_id", nullable = false)
+    private Template template;
 
     @Column(name = "title", nullable = false, length = 200)
     private String title;
