+++ backend/src/main/java/com/hotpack/krocs/domain/templates/dto/TemplateDTO.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.templates.dto;
+
+public class TemplateDTO {
+}
