+++ backend/src/main/java/com/hotpack/krocs/domain/templates/exception/TemplateExceptionType.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.templates.exception;
+
+public class TemplateExceptionType {
+}
