+++ backend/src/main/java/com/hotpack/krocs/domain/templates/repository/SubTemplateRepository.java
@@ -1,0 +1,10 @@
+package com.hotpack.krocs.domain.templates.repository;
+
+import com.hotpack.krocs.domain.templates.domain.SubTemplate;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SubTemplateRepository extends JpaRepository<SubTemplate, Long> {
+
+}
