+++ backend/src/main/java/com/hotpack/krocs/domain/templates/service/TemplateService.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.templates.service;
+
+public interface TemplateService {
+}
