+++ backend/src/main/java/com/hotpack/krocs/domain/templates/service/TemplateServiceImpl.java
@@ -1,0 +1,4 @@
+package com.hotpack.krocs.domain.templates.service;
+
+public class TemplateServiceImpl {
+}
